Align machine identifiers with database schema

+++ backend/src/controlmat.Domain/Interfaces/IMachineRepository.cs
@@ -7,7 +7,7 @@ namespace Controlmat.Domain.Interfaces;
 
 public interface IMachineRepository
 {
-    Task<Machine?> GetByIdAsync(int machineId);
+    Task<Machine?> GetByIdAsync(short machineId);
     Task<IEnumerable<Machine>> GetAllAsync();
 
 }
